Refactor: Update basehub and Next.js dependencies
This commit updates the basehub library and related dependencies to their latest versions. It also upgrades Next.js and React to newer versions, ensuring compatibility and leveraging the latest features.
