+++ basehub.config.ts
@@ -1,0 +1,3 @@
+import { setGlobalConfig } from 'basehub';
+
+setGlobalConfig({})
