+++ package.json
@@ -15,17 +15,17 @@
     "@radix-ui/react-slot": "^1.1.2",
     "@radix-ui/themes": "^3.2.1",
     "@types/js-cookie": "^3.0.6",
-    "basehub": "8.1.19",
+    "basehub": "9.5.2",
     "change-case": "^5.4.4",
     "clsx": "^2.1.1",
     "date-fns": "^3.6.0",
     "js-cookie": "^3.0.5",
     "lucide-react": "^0.378.0",
     "modern-normalize": "^2.0.0",
-    "next": "~15.1.4",
+    "next": "16.0.0",
     "next-themes": "^0.3.0",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
+    "react": "19.2.0",
+    "react-dom": "19.2.0",
     "react-medium-image-zoom": "^5.2.4",
     "react-use-intercom": "^5.4.0",
     "sass": "^1.85.1"
