+++ tsconfig.json
@@ -36,7 +36,7 @@
     "**/*.ts",
     "**/*.tsx",
     ".next/types/**/*.ts"
-  ],
+  , ".basehub/basehub-types.d.ts", "basehub.config.ts"],
   "exclude": [
     "node_modules"
   ]
